+++ ssa.h
@@ -181,7 +181,7 @@ public:
     int id;
     ScriptVariant constValue;
     Constant(ScriptVariant val);
-    Constant(LONG intVal);
+    Constant(s32 intVal);
     // Constant(double dblVal);
     // Constant(const char *strVal);
     virtual bool isConstant();
@@ -447,7 +447,7 @@ public:
     // these make an instruction, insert it at the end of the current block, and return it
     RValue *mkUnaryOp(OpCode op, RValue *src);
     RValue *mkBinaryOp(OpCode op, RValue *src0, RValue *src1);
-    Constant *mkConstInt(LONG val);
+    Constant *mkConstInt(s32 val);
     Constant *mkConstString(char *val);
     Constant *mkConstFloat(double val);
     Constant *mkNull();
